modify makefile
